project_data/django/requirements.txt
asgiref==3.2.10
backcall==0.2.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
configparser==5.0.0
cryptography==3.0
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
mysqlclient==2.0.1
paramiko==2.7.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
uWSGI==2.0.19.1
wcwidth==0.2.5
whitenoise==5.1.0
zipp==3.1.0
